+++ api/types.go
@@ -8,22 +8,22 @@ type CreateTokenRequest struct {
 }
 
 type GetTokenStatusResponse struct {
-	AtBlock  uint64 `json:"atBlock"`
-	Synced   bool   `json:"synced"`
-	Progress uint64 `json:"progress"`
+	AtBlock  int64 `json:"atBlock"`
+	Synced   bool  `json:"synced"`
+	Progress int   `json:"progress"`
 }
 
 type GetTokenResponse struct {
 	ID              string                  `json:"id"`
 	Type            string                  `json:"type"`
-	Decimals        uint64                  `json:"decimals"`
-	StartBlock      uint64                  `json:"startBlock"`
+	Decimals        int                     `json:"decimals"`
+	StartBlock      int64                   `json:"startBlock"`
 	Symbol          string                  `json:"symbol"`
 	TotalSupply     string                  `json:"totalSupply"`
 	Name            string                  `json:"name"`
 	Status          *GetTokenStatusResponse `json:"status"`
-	Size            uint32                  `json:"size"`
-	DefaultStrategy uint64                  `json:"defaultStrategy,omitempty"`
+	Size            int                     `json:"size"`
+	DefaultStrategy int                     `json:"defaultStrategy,omitempty"`
 	Tag             string                  `json:"tag,omitempty"`
 	ChainID         int64                   `json:"chainID"`
 }
@@ -31,7 +31,7 @@ type GetTokenResponse struct {
 type GetTokensItem struct {
 	ID         string `json:"id"`
 	Type       string `json:"type"`
-	StartBlock uint64 `json:"startBlock"`
+	StartBlock int64  `json:"startBlock"`
 	Name       string `json:"name"`
 	Symbol     string `json:"symbol"`
 	Tag        string `json:"tag,omitempty"`
@@ -51,9 +51,9 @@ type TokenHoldersResponse struct {
 }
 
 type CreateCensusResquest struct {
-	StrategyID  uint64 `json:"strategyId"`
-	BlockNumber uint64 `json:"blockNumber"`
-	Anonymous   bool   `json:"anonymous"`
+	StrategyID  int   `json:"strategyId"`
+	BlockNumber int64 `json:"blockNumber"`
+	Anonymous   bool  `json:"anonymous"`
 }
 
 type CreateCensusResponse struct {
@@ -61,21 +61,21 @@ type CreateCensusResponse struct {
 }
 
 type GetCensusResponse struct {
-	CensusID   uint64 `json:"censusId"`
-	StrategyID uint64 `json:"strategyId"`
+	CensusID   int    `json:"censusId"`
+	StrategyID int    `json:"strategyId"`
 	MerkleRoot string `json:"merkleRoot"`
 	URI        string `json:"uri"`
-	Size       int32  `json:"size"`
+	Size       int    `json:"size"`
 	Weight     string `json:"weight"`
 	Anonymous  bool   `json:"anonymous"`
 }
 
 type GetCensusesResponse struct {
-	Censuses []uint64 `json:"censuses"`
+	Censuses []int `json:"censuses"`
 }
 
 type GetStrategiesResponse struct {
-	Strategies []uint64 `json:"strategies"`
+	Strategies []int `json:"strategies"`
 }
 
 type GetStrategyToken struct {
@@ -86,7 +86,7 @@ type GetStrategyToken struct {
 }
 
 type GetStrategyResponse struct {
-	ID        uint64             `json:"id"`
+	ID        int                `json:"id"`
 	Tokens    []GetStrategyToken `json:"tokens"`
 	Predicate string             `json:"strategy"`
 }
